Use MenuLink instead of MenuLinks
